pyproject.toml
[project]
name = "vigil-ai-cli"
version = "0.1.0"
description = "Security scanner for AI-generated code"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
keywords = ["security", "ai", "linter", "sast", "slopsquatting"]

dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "httpx>=0.27",
    "structlog>=24.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
vigil = "vigil.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]
